AlgoQuest_/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.13.1
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.5
